csocket => fixed bug (better: missing code) => handle short writes in sendto() in a correct manner, i.e. avoud out-of-sync problems on stream sockets

+++ src/rofl/common/csocket.cc
@@ -654,7 +654,7 @@ csocket::dequeue_packet()
 					throw eSysCall("sendto");
 				}
 			}
-			else if ((((unsigned int)rc + entry.msg_bytes_sent) < (int)entry.mem->memlen())) {
+			else if (((unsigned int)(rc + entry.msg_bytes_sent) < entry.mem->memlen())) {
 				logging::warn << "[rofl][csocket] short write on socket descriptor:" << sd << std::endl;
 				if (SOCK_STREAM == type) {
 					entry.msg_bytes_sent += rc;
